fix: pii sharing and app config fixes
