+++ src/css/FormSwitchGroup.css
@@ -1,0 +1,7 @@
+.helper-text {
+  font-size: 1.125rem !important;
+}
+
+.z-index-3 {
+  z-index: 3;
+}
