+++ src/css/proctoredExamSettings.css
@@ -1,0 +1,9 @@
+legend {
+  font-size: 1.0rem;
+  font-weight: 700;
+}
+
+select {
+  -webkit-appearance: none;
+  -moz-appearance: none;
+}
