+++ src/pages-and-resources/css/AppSettingsModal.css
@@ -1,0 +1,4 @@
+.modal-footer {
+  position: absolute;
+  width: 100%;
+}
