+++ src/pages-and-resources/css/PageCard.css
@@ -1,0 +1,9 @@
+.desktop-card {
+  width: 19rem;
+  height: 14rem;
+}
+
+.mobile-card {
+  width: 100%;
+  height: 14rem;
+}
