+++ src/pages-and-resources/discussions/app-config-form/apps/shared/AnonymousPostingFields.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { injectIntl, intlShape } from '@edx/frontend-platform/i18n';
 import { TransitionReplace } from '@edx/paragon';
 import FormSwitchGroup from '../../../../../generic/FormSwitchGroup';
-import messages from './messages';
+import messages from '../../messages';
 import AppConfigFormDivider from './AppConfigFormDivider';
 
 function AnonymousPostingFields({
