+++ src/pages-and-resources/discussions/css/FeaturesTable.css
@@ -1,0 +1,4 @@
+table {
+  font-size: 14px;
+  color: black;
+}
